Compact InputSection from 80 to 72 bytes
Continue narrowing the highest-count struct toward mold-rust's 64-byte
InputSection:

  - input_addr u64 -> u32. It is a subsection's address within its
    object's address space; object files stay far under 4 GiB. The
    find_subsec search and the staging arithmetic cast it up to u64
    where they mix it with 64-bit addresses.
  - p2align u32 -> u8. An alignment is a small power of two; a byte is
    ample. Output-section alignment (ChunkHeader.p2align, still u32)
    takes the max with a widening cast.

Another 8 bytes off the ~4.4M-element arena on a debug link. size stays
u64 and addr stays u64 (an arm64 image's addresses sit above the 4 GiB
__PAGEZERO). size asserted at 72. 322 integration tests pass.

Note: not re-benchmarked - the scratchpad debug-link inputs were
cleared - but this is the same type-narrowing as the 112->96->88->80
steps, each of which measured a few ms on clang-dbg.

+++ src/input_files.rs
@@ -73,14 +73,14 @@ pub fn find_subsec(
     subsecs: &[usize],
     addr: u64,
 ) -> Option<(usize, u64)> {
-    let i = subsecs.partition_point(|&id| isecs[id].input_addr <= addr);
+    let i = subsecs.partition_point(|&id| isecs[id].input_addr as u64 <= addr);
     if i == 0 {
         return None;
     }
     let id = subsecs[i - 1];
     let isec = &isecs[id];
-    if addr < isec.input_addr + isec.size || (isec.size == 0 && addr == isec.input_addr) {
-        Some((id, addr - isec.input_addr))
+    if addr < isec.input_addr as u64 + isec.size || (isec.size == 0 && addr == isec.input_addr as u64) {
+        Some((id, addr - isec.input_addr as u64))
     } else {
         None
     }
@@ -367,8 +367,8 @@ pub fn stage_object<E: Arch>(
             isecs.push(InputSection {
                 obj: u32::MAX,
                 hdr: sect,
-                p2align: sect.p2align,
-                input_addr: start,
+                p2align: sect.p2align as u8,
+                input_addr: start as u32,
                 size: end - start,
                 data_ptr: if contents.is_empty() { 0 } else { contents.as_ptr() as usize },
                 rel_offset: 0,
@@ -418,7 +418,7 @@ pub fn stage_object<E: Arch>(
 
         let mut pos = 0;
         for &sub in &by_ordinal[i] {
-            let sub_off = (isecs[sub].input_addr - sect.addr) as u32;
+            let sub_off = (isecs[sub].input_addr as u64 - sect.addr) as u32;
             let end = sub_off + isecs[sub].size as u32;
             let start = obj_relocs.len();
             while pos < rels.len() && rels[pos].offset < end {
@@ -987,7 +987,7 @@ fn parse_compact_unwind<E: Arch>(
 ) {
     let geo: Vec<(u64, u64, usize)> = subsecs
         .iter()
-        .map(|&id| (isecs[id].input_addr, isecs[id].size, id))
+        .map(|&id| (isecs[id].input_addr as u64, isecs[id].size, id))
         .collect();
     let find_subsec = |addr: u64| -> Option<(usize, u32)> {
         let i = geo.partition_point(|&(start, _, _)| start <= addr);
@@ -1163,7 +1163,7 @@ fn parse_eh_frame<E: Arch>(
 ) {
     let geo: Vec<(u64, u64, usize)> = subsecs
         .iter()
-        .map(|&id| (isecs[id].input_addr, isecs[id].size, id))
+        .map(|&id| (isecs[id].input_addr as u64, isecs[id].size, id))
         .collect();
     let find_local = |addr: u64| -> Option<(usize, u32)> {
         let i = geo.partition_point(|&(start, _, _)| start <= addr);
